+++ crates/recourse/src/http/policy/authentication/mod.rs
@@ -2,7 +2,10 @@
 
 mod basic;
 mod bearer;
+mod challenge;
+mod grammar;
 mod realm;
+mod response;
 
 pub use basic::{BasicChallenge, BasicChallengeError, BasicUnauthorized};
 pub use bearer::{BearerChallenge, BearerChallengeError, BearerUnauthorized};